try to control dtypes in metrics

+++ src/condor_tensorflow/metrics.py
@@ -124,7 +124,8 @@ class OrdinalEarthMoversDistance(tf.keras.metrics.Metric):
     # remove all dimensions of size 1 (e.g., from [[1], [2]], to [1, 2])
     #y_true = tf.squeeze(y_true)
 
-    y_dist = tf.map_fn(fn=lambda y: tf.abs(y-tf.range(num_classes)),
+    y_dist = tf.map_fn(fn=lambda y:
+                       tf.abs(y-tf.range(num_classes,dtype=y_pred.dtype)),
                        elems=y_true)
 
     self.emds.assign_add(tf.reduce_sum(tf.math.multiply(y_dist,cum_probs)))
@@ -171,7 +172,8 @@ class SparseOrdinalEarthMoversDistance(OrdinalEarthMoversDistance):
     # remove all dimensions of size 1 (e.g., from [[1], [2]], to [1, 2])
     #y_true = tf.squeeze(y_true)
 
-    y_dist = tf.map_fn(fn=lambda y: tf.abs(y-tf.range(num_classes)),
+    y_dist = tf.map_fn(fn=lambda y:
+                       tf.abs(y-tf.range(num_classes,dtype=y_pred.dtype)),
                        elems=y_true)
 
     self.emds.assign_add(tf.reduce_sum(tf.math.multiply(y_dist,cum_probs)))
